% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.to.granges.R
\name{bed.to.granges}
\alias{bed.to.granges}
\title{Convert BED file to GRanges object.}
\usage{
bed.to.granges(bed.file)
}
\arguments{
\item{bed.file}{BED file to be converted.}
}
\value{
GRanges object.
}
\description{
Convert BED file to GRanges object.
}
