% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.to.grangeslist.R
\name{bed.to.grangeslist}
\alias{bed.to.grangeslist}
\title{Converts bed file to grangeslist.}
\usage{
bed.to.grangeslist(fname)
}
\arguments{
\item{fname}{Name of bed file.}
}
\value{
GRangeslist of the input file.
}
\description{
Converts bed file to grangeslist.
}
