% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.continuous.R
\name{bin.continuous}
\alias{bin.continuous}
\title{Divide continuous feature scores into specified number of bins.}
\usage{
bin.continuous(feature.name, feature.url, nbins, genome.build = "Ch37")
}
\arguments{
\item{feature.name}{Continuous feature name.}

\item{feature.url}{URL of continuous feature.}

\item{nbins}{Number of bins to divide continuous feature score into.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Binned continuous feature scores.
}
\description{
Divide continuous feature scores into specified number of bins.
}
