% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedMax.R
\name{binnedMax}
\alias{binnedMax}
\title{Calculate maximum score in each bin.}
\usage{
binnedMax(bins, numvar, mcolname)
}
\arguments{
\item{bins}{Genomic bins GRanges.}

\item{numvar}{Feature scores across genome in RleList.}

\item{mcolname}{Colname of feature score, default = "score".}
}
\value{
Maximum score for each bin.
}
\description{
Calculate maximum score in each bin.
}
