% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.lasso.input.R
\name{create.lasso.input}
\alias{create.lasso.input}
\title{Create covariates and response matrices as input for LASSO.}
\usage{
create.lasso.input(sampled.sites.file, genomic.features.file, mutation.type,
  cores = 1)
}
\arguments{
\item{sampled.sites.file}{Sampled mutated and non-mutated sites RDS file.}

\item{genomic.features.file}{Text file containing URLs of potential continuous and discrete epigenetic features to select from.}

\item{mutation.type}{Type of mutation, snv or indel.}

\item{cores}{Number of cores, default = 1.}
}
\value{
List containing covariate matrix, response matrix, final continuous feature urls, final discrete feature urls.
}
\description{
Create covariates and response matrices as input for LASSO.
}
