% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution.test.R
\name{dilution.test}
\alias{dilution.test}
\title{Dilution analysis.}
\usage{
dilution.test(
  snv.mutations.file,
  mask.regions.file = system.file("extdata", "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package = "MutSpot"),
  cores = 1,
  cutoff.nucleotide = 0.9,
  cutoff.nucleotide.new = NULL,
  cutoff.features = 0.75,
  cutoff.features.new.snv = NULL,
  genomic.features.snv = NULL,
  genomic.features.indel = NULL,
  genomic.features = NULL,
  genomic.features.fixed.snv = NULL,
  genomic.features.fixed.indel = NULL,
  genomic.features.fixed = NULL,
  feature.dir,
  mutCovariate.table.snv.file,
  mutCovariate.count.snv.file,
  continuous.features.selected.snv.url.file,
  discrete.features.selected.snv.url.file,
  nucleotide.selected.file,
  sample.specific.features.url.file = NULL,
  output.dir,
  genome.build = "Ch37"
)
}
\arguments{
\item{snv.mutations.file}{SNV mutations MAF file.}

\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, depends on genome build, default file = mask_regions.RDS, Ch37.}

\item{all.sites.file}{All sites in whole genome RDS file, depends on genome build, default file = all_sites.RDS, Ch37.}

\item{cores}{Number of cores, default = 1.}

\item{cutoff.nucleotide}{Frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1, default = 0.90.}

\item{cutoff.nucleotide.new}{Updated frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1.}

\item{cutoff.features}{Frequency cutoff/threshold to determine epigenetic features used in prediction model, ranges from 0.5 to 1, default = 0.75.}

\item{cutoff.features.new.snv}{Updated frequency cutoff/threshold to determine SNV epigenetic features used in prediction model, ranges from 0.5 to 1.}

\item{genomic.features.snv}{Text file containing URLs of potential continuous and discrete SNV epigenetic features to select from, default = NULL.}

\item{genomic.features.indel}{Text file containing URLs of potential continuous and discrete indel epigenetic features to select from, default = NULL.}

\item{genomic.features}{Text file containing URLs of potential continuous and discrete SNV and indel epigenetic features to select from, default = NULL.}

\item{genomic.features.fixed.snv}{Text file containing URLs of fixed continuous and discrete SNV epigenetic features, default = NULL.}

\item{genomic.features.fixed.indel}{Text file containing URLs of fixed continuous and discrete indel epigenetic features, default = NULL.}

\item{genomic.features.fixed}{Text file containing URLs of fixed continuous and discrete SNV and indel epigenetic features, default = NULL.}

\item{feature.dir}{Directory containing binned feature bed files.}

\item{mutCovariate.table.snv.file}{RDS file, SNV covariates sparse matrix.}

\item{mutCovariate.count.snv.file}{RDS file, SNV response matrix.}

\item{continuous.features.selected.snv.url.file}{Text file containing URLs of SNV continuous features selected for model.}

\item{discrete.features.selected.snv.url.file}{Text file containing URLs of SNV discrete features selected for model.}

\item{nucleotide.selected.file}{Nucleotide context selected for model RDS file.}

\item{sample.specific.features.url.file}{Text file containing sample specific SNV features, default = NULL.}

\item{output.dir}{Save plots in given output directory.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
McFadden's R2 for different dilution size.
}
\description{
Dilution analysis.
}
