% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenetic.selection.R
\name{epigenetic.selection}
\alias{epigenetic.selection}
\title{Select epigentic features through LASSO.}
\usage{
epigenetic.selection(sampled.sites.snv.file, sampled.sites.indel.file,
  genomic.features.snv = NULL, genomic.features.indel = NULL,
  genomic.features = NULL, genomic.features.fixed.snv = NULL,
  genomic.features.fixed.indel = NULL, genomic.features.fixed = NULL,
  cores = 1, cutoff = 0.75)
}
\arguments{
\item{sampled.sites.snv.file}{Sampled SNV mutated and non-mutated sites RDS file.}

\item{sampled.sites.indel.file}{Sampled indel mutated and non-mutated sites RDS file.}

\item{genomic.features.snv}{Text file containing URLs of potential continuous and discrete SNV epigenetic features to select from, default = NULL.}

\item{genomic.features.indel}{Text file containing URLs of potential continuous and discrete indel epigenetic features to select from, default = NULL.}

\item{genomic.features}{Text file containing URLs of potential continuous and discrete SNV and indel epigenetic features to select from, default = NULL.}

\item{genomic.features.fixed.snv}{Text file containing URLs of fixed continuous and discrete SNV epigenetic features, default = NULL.}

\item{genomic.features.fixed.indel}{Text file containing URLs of fixed continuous and discrete indel epigenetic features, default = NULL.}

\item{genomic.features.fixed}{Text file containing URLs of fixed continuous and discrete SNV and indel epigenetic features, default = NULL.}

\item{cores}{Number of cores, default = 1.}

\item{cutoff}{Frequency cutoff/threshold to determine epigenetic features used in prediction model, ranges from 0.5 to 1, default = 0.75.}
}
\value{
A list containing frequency of SNV epigenetic features selected through lasso, SNV continuous and discrete features that passed the threshold, frequency of indel epigenetic features selected through lasso, indel continuous and discrete features that passed the threshold.
}
\description{
Select epigentic features through LASSO.
}
