% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.fit.nucleotide.R
\name{lasso.fit.nucleotide}
\alias{lasso.fit.nucleotide}
\title{1 iteration of LASSO selection on a subset of data.}
\usage{
lasso.fit.nucleotide(x_data, y_data, s)
}
\arguments{
\item{x_data}{Covariates matrix.}

\item{y_data}{Response matrix.}

\item{s}{Seed for random rumber.}
}
\value{
Nucleotide contexts selected for 1 iteration of LASSO.
}
\description{
1 iteration of LASSO selection on a subset of data.
}
