% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.mutrate.R
\name{local.mutrate}
\alias{local.mutrate}
\title{Calculate local mutation rate.}
\usage{
local.mutrate(snv.mutations.file, indel.mutations.file)
}
\arguments{
\item{snv.mutations.file}{SNV mutations MAF file.}

\item{indel.mutations.file}{Indel mutations MAF file.}
}
\value{
A list containing binned SNV local mutation rates and binned indel local mutation rates.
}
\description{
Calculate local mutation rate.
}
