% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.indel.freq.table.genome.R
\name{mutCovariate.indel.freq.table.genome}
\alias{mutCovariate.indel.freq.table.genome}
\title{Prepare covariates matrix for all sites in whole genome/specified region.}
\usage{
mutCovariate.indel.freq.table.genome(continuous.features, discrete.features,
  polyAT, polyCG, sites)
}
\arguments{
\item{continuous.features}{Continuous epigenetic features selected for model fitting.}

\item{discrete.features}{Discrete epigenetic features selected for model fitting.}

\item{polyAT}{All polyA or polyT positions in whole genome.}

\item{polyCG}{All polyC or polyG positions in whole genome.}

\item{sites}{All sites in whole genome/specified region.}
}
\value{
Covariates matrix for all sites in whole genome/specified region.
}
\description{
Prepare covariates matrix for all sites in whole genome/specified region.
}
