% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.indel.freq.table.muts.R
\name{mutCovariate.indel.freq.table.muts}
\alias{mutCovariate.indel.freq.table.muts}
\title{Prepare covariates matrix for mutated sites in a specfic chromosome.}
\usage{
mutCovariate.indel.freq.table.muts(
  continuous.features,
  discrete.features,
  sample.specific.features,
  polyAT,
  polyCG,
  sites
)
}
\arguments{
\item{continuous.features}{Continuous epigenetic features selected for model fitting.}

\item{discrete.features}{Discrete epigenetic features selected for model fitting.}

\item{sample.specific.features}{Sample-specific features.}

\item{polyAT}{All polyA or polyT positions in whole genome.}

\item{polyCG}{All polyC or polyG positions in whole genome.}

\item{sites}{Mutated sites in a specific chromosome.}
}
\value{
Covariate matrix for mutated sites.
}
\description{
Prepare covariates matrix for mutated sites in a specfic chromosome.
}
