% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.indel.R
\name{mutCovariate.indel}
\alias{mutCovariate.indel}
\title{Prepare covariates matrix by chromosome for indel model fitting.}
\usage{
mutCovariate.indel(mask.regions.file = system.file("extdata",
  "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package =
  "MutSpot"), continuous.features.selected.indel.url.file,
  discrete.features.selected.indel.url.file,
  sample.specific.features.url.file = NULL, indel.mutations.file,
  region.of.interest = NULL, indel.mutations.file2, cores = 1,
  chr.interest = "chr1")
}
\arguments{
\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, default file = mask_regions.RDS.}

\item{all.sites.file}{All sites in whole genome RDS file, default file = all_sites.RDS.}

\item{continuous.features.selected.indel.url.file}{Text file containing URLs of indel continuous features selected for model.}

\item{discrete.features.selected.indel.url.file}{Text file containing URLs of indel discrete features selected for model.}

\item{sample.specific.features.url.file}{Text file containing sample specific indel features, default = NULL.}

\item{indel.mutations.file}{Indel mutations found in region of interest MAF file.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{indel.mutations.file2}{Indel mutations MAF file.}

\item{cores}{Number of cores, default = 1.}

\item{chr.interest}{Chromosome of interest, default = "chr1".}
}
\value{
Saves covariates matrices for mutated sites and whole genome/specified region by chromosome.
}
\description{
Prepare covariates matrix by chromosome for indel model fitting.
}
