% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.snv.compile.R
\name{mutCovariate.snv.compile}
\alias{mutCovariate.snv.compile}
\title{Compile covariates matrix for all chromosomes together.}
\usage{
mutCovariate.snv.compile(mask.regions.file = system.file("extdata",
  "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package =
  "MutSpot"), snv.mutations.file, sample.specific.features.url.file = NULL,
  region.of.interest, cores = 1, snv.mutations.file2)
}
\arguments{
\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, default file = mask_regions.RDS.}

\item{all.sites.file}{All sites in whole genome RDS file, default file = all_sites.RDS.}

\item{snv.mutations.file}{SNV mutations found in region of interest MAF file.}

\item{sample.specific.features.url.file}{Text file containing sample specific SNV features, default = NULL.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{cores}{Number of cores, default = 1.}

\item{snv.mutations.file2}{SNV mutations MAF file.}
}
\value{
Full covariates matrix for all chromosomes together.
}
\description{
Compile covariates matrix for all chromosomes together.
}
