% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.snv.freq.table.muts.R
\name{mutCovariate.snv.freq.table.muts}
\alias{mutCovariate.snv.freq.table.muts}
\title{Prepare covariates matrix for mutated sites in a specfic chromosome.}
\usage{
mutCovariate.snv.freq.table.muts(
  continuous.features,
  discrete.features,
  precompute.motif.pos,
  nucleotide.selected,
  sample.specific.features,
  sites
)
}
\arguments{
\item{continuous.features}{Continuous epigenetic features selected for model fitting.}

\item{discrete.features}{Discrete epigenetic features selected for model fitting.}

\item{precompute.motif.pos}{All selected nucleotide contexts' positions in a specific chromosome.}

\item{nucleotide.selected}{Nucleotide contexts selected for model fitting.}

\item{sample.specific.features}{Sample-specific features.}

\item{sites}{Mutated sites in a specific chromosome.}
}
\value{
Covariate matrix for mutated sites.
}
\description{
Prepare covariates matrix for mutated sites in a specfic chromosome.
}
