% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCovariate.snv.R
\name{mutCovariate.snv}
\alias{mutCovariate.snv}
\title{Prepare covariates matrix by chromosome for SNV model fitting.}
\usage{
mutCovariate.snv(mask.regions.file = system.file("extdata",
  "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package =
  "MutSpot"), nucleotide.selected.file,
  continuous.features.selected.snv.url.file,
  discrete.features.selected.snv.url.file,
  sample.specific.features.url.file = NULL, snv.mutations.file,
  region.of.interest = NULL, snv.mutations.file2, cores = 1,
  chr.interest = "chr1")
}
\arguments{
\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, default file = mask_regions.RDS.}

\item{all.sites.file}{All sites in whole genome RDS file, default file = all_sites.RDS.}

\item{nucleotide.selected.file}{Nucleotide context selected for model RDS file.}

\item{continuous.features.selected.snv.url.file}{Text file containing URLs of SNV continuous features selected for model.}

\item{discrete.features.selected.snv.url.file}{Text file containing URLs of SNV discrete features selected for model.}

\item{sample.specific.features.url.file}{Text file sample specific SNV features, default = NULL.}

\item{snv.mutations.file}{SNV mutations found in region of interest MAF file.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{snv.mutations.file2}{SNV mutations MAF file.}

\item{cores}{Number of cores, default = 1.}

\item{chr.interest}{Chromosome of interest, default = "chr1".}
}
\value{
Saves covariates matrices for mutated sites and whole genome/specified region by chromosome.
}
\description{
Prepare covariates matrix by chromosome for SNV model fitting.
}
