% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutLRFit.indel.R
\name{mutLRFit.indel}
\alias{mutLRFit.indel}
\title{Fit logistic regression prediction model for indel.}
\usage{
mutLRFit.indel(mutCovariate.table.indel.file, mutCovariate.count.indel.file,
  continuous.features.selected.indel.url.file,
  discrete.features.selected.indel.url.file,
  sample.specific.features.url.file = NULL, fit.sparse = FALSE,
  drop = FALSE)
}
\arguments{
\item{mutCovariate.table.indel.file}{RDS file, indel covariates sparse matrix.}

\item{mutCovariate.count.indel.file}{RDS file, indel response matrix.}

\item{continuous.features.selected.indel.url.file}{Text file containing URLs of indel continuous features selected for model.}

\item{discrete.features.selected.indel.url.file}{Text file containing URLs of indel discrete features selected for model.}

\item{sample.specific.features.url.file}{Text file containing sample specific indel features, default = NULL.}

\item{fit.sparse}{To fit model using glmnet or glm, default = FALSE.}

\item{drop}{To drop insignificant features from fitted model or not, default = FALSE.}
}
\value{
Fitted indel prediction model.
}
\description{
Fit logistic regression prediction model for indel.
}
