% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutLRFit.snv.R
\name{mutLRFit.snv}
\alias{mutLRFit.snv}
\title{Fit logistic regression prediction model for SNV.}
\usage{
mutLRFit.snv(
  mutCovariate.table.snv.file,
  mutCovariate.count.snv.file,
  continuous.features.selected.snv.url.file,
  discrete.features.selected.snv.url.file,
  nucleotide.selected.file,
  sample.specific.features.url.file = NULL,
  fit.sparse = FALSE,
  drop = FALSE,
  output.dir
)
}
\arguments{
\item{mutCovariate.table.snv.file}{RDS file, SNV covariates sparse matrix.}

\item{mutCovariate.count.snv.file}{RDS file, SNV response matrix.}

\item{continuous.features.selected.snv.url.file}{Text file containing URLs of SNV continuous features selected for model.}

\item{discrete.features.selected.snv.url.file}{Text file containing URLs of SNV discrete features selected for model.}

\item{nucleotide.selected.file}{Nucleotide context selected for model RDS file.}

\item{sample.specific.features.url.file}{Text file containing sample specific SNV features, default = NULL.}

\item{fit.sparse}{To fit model using glmnet or glm, default = FALSE.}

\item{drop}{To drop insignificant features from fitted model or not, default = FALSE.}

\item{output.dir}{Save plot in given output directory.}
}
\value{
Fitted SNV prediction model.
}
\description{
Fit logistic regression prediction model for SNV.
}
