% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.indel.find.polymer.R
\name{mutPredict.indel.find.polymer}
\alias{mutPredict.indel.find.polymer}
\title{Extract polyA/T/C/G for each site in indel hotspots.}
\usage{
mutPredict.indel.find.polymer(seq)
}
\arguments{
\item{seq}{Dataframe containing DNA sequences of each site in the indel hotspot.}
}
\value{
PolyA/T/C/G feature matrix.
}
\description{
Extract polyA/T/C/G for each site in indel hotspots.
}
