% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.indel.prepare.R
\name{mutPredict.indel.prepare}
\alias{mutPredict.indel.prepare}
\title{Mutation hotspot recurrence prediction for indel.}
\usage{
mutPredict.indel.prepare(
  mask.regions.file = system.file("extdata", "mask_regions.RDS", package = "MutSpot"),
  continuous.features.selected.indel.url.file,
  discrete.features.selected.indel.url.file,
  sample.specific.features.url.file = NULL,
  indel.mutations.file,
  indel.mutations.file2,
  indel.model.file,
  collapse.regions = FALSE,
  region.of.interest,
  cores = 1,
  min.count = 2,
  hotspot.size = 21,
  genome.size = 2533374732,
  hotspots = TRUE,
  output.dir,
  genome.build = "Ch37"
)
}
\arguments{
\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, depends on genome build, default file = mask_regions.RDS, Ch37.}

\item{continuous.features.selected.indel.url.file}{Text file containing URLs of indel continuous features selected for model.}

\item{discrete.features.selected.indel.url.file}{Text file containing URLs of indel discrete features selected for model.}

\item{sample.specific.features.url.file}{Text file containing URLs of sample specific features, default = NULL.}

\item{indel.mutations.file}{Indel mutations found in region of interest MAF file.}

\item{indel.mutations.file2}{Indel mutations MAF file.}

\item{indel.model.file}{Indel model.}

\item{collapse.regions}{To collapse region of interest or not, default = FALSE.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{cores}{Number of cores, default = 1.}

\item{min.count}{Minimum number of mutated samples in each hotspot, default = 2.}

\item{hotspot.size}{Size of each hotspot, default = 21.}

\item{genome.size}{Genome size, depends on genome build, default = 2533374732, Ch37.}

\item{hotspots}{To run hotspot analysis or region-based analysis, default = TRUE.}

\item{output.dir}{Save plot in given output directory.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Prepare intermediate files for hotspot prediction.
}
\description{
Mutation hotspot recurrence prediction for indel.
}
