% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.indel.run.lr.R
\name{mutPredict.indel.run.lr}
\alias{mutPredict.indel.run.lr}
\title{Run mutation recurrence for indel.}
\usage{
mutPredict.indel.run.lr(roi, maf.indel, maf.indel2, model.indel,
  continuous.features.indel, discrete.features.indel, sample.specific.features,
  continuous.sample.specific, min.count = 2, genome.size = 2533374732,
  cores = 1, only.samples = NULL, debug = FALSE)
}
\arguments{
\item{roi}{GRanges containing all hotspots.}

\item{maf.indel}{GRanges containing indel mutations in region of interest.}

\item{maf.indel2}{GRanges containing all mutations.}

\item{model.indel}{Indel model.}

\item{continuous.features.indel}{Full scores list of indel continuous features selected for model.}

\item{discrete.features.indel}{Full scores list of indel discrete features selected for model.}

\item{sample.specific.features}{Sample-specific features.}

\item{continuous.sample.specific}{Continuous features in the final model.}

\item{min.count}{Minimum number of mutated samples in each hotspot, default = 2.}

\item{genome.size}{Total number of hotspots to run analysis on, default = 2533374732.}

\item{cores}{Number of cores, default = 1.}

\item{only.samples}{Restrict analysis to certain samples, default = NULL.}

\item{debug}{To print errors, default = FALSE.}
}
\value{
Dataframe containing predicted hotspots significance.
}
\description{
Run mutation recurrence for indel.
}
