% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.snv.get.features.R
\name{mutPredict.snv.get.features}
\alias{mutPredict.snv.get.features}
\title{Mutation hotspot recurrence prediction for SNV.}
\usage{
mutPredict.snv.get.features(
  roi,
  continuous.features.snv,
  discrete.features.snv,
  motifs,
  genome.build = "Ch37"
)
}
\arguments{
\item{roi}{Each SNV hotspot in GRanges.}

\item{continuous.features.snv}{Full scores list of SNV continuous features selected for model.}

\item{discrete.features.snv}{Full scores list of SNV discrete features selected for model.}

\item{motifs}{Selected nucleotide contexts to be extracted.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Feature matrix.
}
\description{
Mutation hotspot recurrence prediction for SNV.
}
