% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.snv.prepare.R
\name{mutPredict.snv.prepare}
\alias{mutPredict.snv.prepare}
\title{Run mutation recurrence for SNV.}
\usage{
mutPredict.snv.prepare(mask.regions.file = system.file("extdata",
  "mask_regions.RDS", package = "MutSpot"), nucleotide.selected.file,
  continuous.features.selected.snv.url.file,
  discrete.features.selected.snv.url.file,
  sample.specific.features.url.file = NULL, snv.mutations.file,
  snv.mutations.file2, collapse.regions = FALSE, region.of.interest,
  cores = 1, snv.model.file, min.count = 2, hotspot.size = 21,
  genome.size = 2533374732, hotspots = TRUE, output.dir)
}
\arguments{
\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, default file = mask_regions.RDS.}

\item{nucleotide.selected.file}{Nucleotide context selected for model RDS file.}

\item{continuous.features.selected.snv.url.file}{Text file containing URLs of SNV continuous features selected for model.}

\item{discrete.features.selected.snv.url.file}{Text file containing URLs of SNV discrete features selected for model.}

\item{sample.specific.features.url.file}{Text file containing URLs of sample specific features, default = NULL.}

\item{snv.mutations.file}{SNV mutations found in region of interest MAF file.}

\item{snv.mutations.file2}{SNV mutations MAF file.}

\item{collapse.regions}{To collapse region of interest or not, default = FALSE.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{cores}{Number of cores, default = 1.}

\item{snv.model.file}{SNV model.}

\item{min.count}{Minimum number of mutated samples in each hotspot, default = 2.}

\item{hotspot.size}{Size of each hotspot, default = 21.}

\item{genome.size}{Genome size, default = 2533374732.}

\item{hotspots}{To run hotspot analysis or region-based analysis, default = TRUE.}

\item{output.dir}{Save temporary files in given output directory.}
}
\value{
Prepare intermediate files for hotspot prediction.
}
\description{
Run mutation recurrence for SNV.
}
