% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.snv.run.lr.R
\name{mutPredict.snv.run.lr}
\alias{mutPredict.snv.run.lr}
\title{Run mutation recurrence for SNV.}
\usage{
mutPredict.snv.run.lr(roi, maf.snv, maf.snv2, model.snv,
  continuous.features.snv, discrete.features.snv, motifs, nucleotide.selected,
  sample.specific.features, continuous.sample.specific, min.count = 2,
  genome.size = 2533374732, cores = 1, only.samples = NULL,
  debug = FALSE)
}
\arguments{
\item{roi}{GRanges containing all hotspots.}

\item{maf.snv}{GRanges containing SNV mutations in region of interest.}

\item{maf.snv2}{GRanges containing all SNV mutations.}

\item{model.snv}{SNV model.}

\item{continuous.features.snv}{Full scores list of SNV continuous features selected for model.}

\item{discrete.features.snv}{Full scores list of SNV discrete features selected for model.}

\item{motifs}{Selected nucleotide contexts to be extracted.}

\item{nucleotide.selected}{Selected nucleotide contexts table.}

\item{sample.specific.features}{Sample-specific features.}

\item{continuous.sample.specific}{Continuous features in the final model.}

\item{min.count}{Minimum number of mutated samples in each hotspot, default = 2.}

\item{genome.size}{Total number of hotspots to run analysis on, default = 2533374732.}

\item{cores}{Number of cores, default = 1.}

\item{only.samples}{Restrict analysis to certain samples, default = NULL.}

\item{debug}{To print errors, default = FALSE.}
}
\value{
Dataframe containing predicted hotspots significance.
}
\description{
Run mutation recurrence for SNV.
}
