% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSpot.R
\name{MutSpot}
\alias{MutSpot}
\title{Runs all or selected steps in MutSpot analysis.}
\usage{
MutSpot(
  run.to = c(1:2, 3.1, 3.2, 4.1, 4.2, 5.1, 5.2, 5.3, 5.4, 5.5, 6, 7),
  working.dir = NULL,
  chromosomes = c(1:22, "X"),
  snv.mutations = NULL,
  indel.mutations = NULL,
  mask.regions.file = system.file("extdata", "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package = "MutSpot"),
  region.of.interest = NULL,
  sample = T,
  cores = 1,
  cutoff.nucleotide = 0.9,
  cutoff.nucleotide.new = NULL,
  top.nucleotide = NULL,
  genomic.features.snv = NULL,
  genomic.features.indel = NULL,
  genomic.features = NULL,
  genomic.features.fixed.snv = NULL,
  genomic.features.fixed.indel = NULL,
  genomic.features.fixed = NULL,
  sample.snv.features = NULL,
  sample.indel.features = NULL,
  cutoff.features = 0.75,
  cutoff.features.new.snv = NULL,
  cutoff.features.new.indel = NULL,
  top.features = NULL,
  fit.sparse = FALSE,
  drop = FALSE,
  min.count.snv = 2,
  min.count.indel = 2,
  hotspot.size = 21,
  genome.size = 2533374732,
  hotspots = TRUE,
  collapse.regions = FALSE,
  promoter.file = system.file("extdata", "Ensembl75.promoters.coding.bed", package =
    "MutSpot"),
  utr3.file = system.file("extdata", "Ensembl75.3UTR.coding.bed", package = "MutSpot"),
  utr5.file = system.file("extdata", "Ensembl75.5UTR.coding.bed", package = "MutSpot"),
  other.annotations = NULL,
  fdr.cutoff = 0.05,
  color.line = "red",
  color.dots = "maroon1",
  merge.hotspots = TRUE,
  color.muts = "orange",
  top.no = 3,
  debug = FALSE,
  dilution.analysis = FALSE,
  genome.build = "Ch37"
)
}
\arguments{
\item{run.to}{Numeric vector defining which steps to run, default = 1, 2, 3.1, 3.2, 4.1, 4.2, 5.1, 5.2, 5.3, 5.4, 5.5, 6, 7.}

\item{working.dir}{Working directory, default = NULL will use current working directory.}

\item{chromosomes}{Character vector defining which chromosomes to compute feature matrix on, default = chr1-chrX.}

\item{snv.mutations}{SNV mutations MAF file, default = NULL.}

\item{indel.mutations}{Indel mutations MAF file, default = NULL.}

\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, depends on genome build, default = mask_regions.RDS, Ch37.}

\item{all.sites.file}{All sites in whole genome RDS file, depends on genome build, default file = all_sites.RDS, Ch37.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{sample}{To sample for non-mutated sites or to use all sites in region of interest, default = TRUE.}

\item{cores}{Number of cores, default = 1.}

\item{cutoff.nucleotide}{Frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1, default = 0.90.}

\item{cutoff.nucleotide.new}{Updated frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1.}

\item{top.nucleotide}{Number of top nucleotide contexts to select, default = NULL.}

\item{genomic.features.snv}{Text file containing URLs of potential continuous and discrete SNV epigenetic features to select from, default = NULL.}

\item{genomic.features.indel}{Text file containing URLs of potential continuous and discrete indel epigenetic features to select from, default = NULL.}

\item{genomic.features}{Text file containing URLs of potential continuous and discrete SNV and indel epigenetic features to select from, default = NULL.}

\item{genomic.features.fixed.snv}{Text file containing URLs of fixed continuous and discrete SNV epigenetic features, default = NULL.}

\item{genomic.features.fixed.indel}{Text file containing URLs of fixed continuous and discrete indel epigenetic features, default = NULL.}

\item{genomic.features.fixed}{Text file containing URLs of fixed continuous and discrete SNV and indel epigenetic features, default = NULL.}

\item{sample.snv.features}{Text file containing sample specific SNV features, default = NULL.}

\item{sample.indel.features}{Text file containing sample specific indel features, default = NULL.}

\item{cutoff.features}{Frequency cutoff/threshold to determine epigenetic features used in prediction model, ranges from 0.5 to 1, default = 0.75.}

\item{cutoff.features.new.snv}{Updated frequency cutoff/threshold to determine SNV epigenetic features used in prediction model, ranges from 0.5 to 1.}

\item{cutoff.features.new.indel}{Updated frequency cutoff/threshold to determine indel epigenetic features used in prediction model, ranges from 0.5 to 1.}

\item{top.features}{Number of top genomic features to select, default = NULL.}

\item{fit.sparse}{To fit model using glmnet or glm, default = FALSE.}

\item{drop}{To drop insignificant features from fitted model or not, default = FALSE.}

\item{min.count.snv}{Minimum number of mutated samples in each SNV hotspot, default = 2.}

\item{min.count.indel}{Minimum number of mutated samples in each indel hotspot, default = 2.}

\item{hotspot.size}{Size of each hotspot, default = 21.}

\item{genome.size}{Genome size, depends on genome build, default = 2533374732, Ch37.}

\item{hotspots}{To run hotspot analysis or region-based analysis, default = TRUE.}

\item{collapse.regions}{To collapse region of interest or not, default = FALSE.}

\item{promoter.file}{Promoter regions bed file, depends on genome build, default file = Ensembl75.promoters.coding.bed, Ch37.}

\item{utr3.file}{3'UTR regions bed file, depends on genome.build, default file = Ensembl75.3UTR.coding.bed, Ch37.}

\item{utr5.file}{5'UTR regions bed file, depends on genome build, default file = Ensembl75.5UTR.coding.bed, Ch37.}

\item{other.annotations}{Text file containing URLs of additional regions to be annotated, default = NULL.}

\item{fdr.cutoff}{FDR cutoff, default = 0.05.}

\item{color.line}{Color given FDR cutoff, default = red.}

\item{color.dots}{Color hotspots that passed given FDR cutoff, default = maroon1.}

\item{merge.hotspots}{To plot overlapping hotspots as 1 hotspot or individual hotspots, default = TRUE.}

\item{color.muts}{Color points, default = orange.}

\item{top.no}{Number of top hotspots to plot, default = 3.}

\item{debug}{To keep intermediate output files or not, default = FALSE.}

\item{dilution.analysis}{To run dilution test or not, default = FALSE.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Corresponding output from each step in MutSpot analysis.
}
\description{
Runs all or selected steps in MutSpot analysis.
}
