% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot feature importance.}
\usage{
plot_feature_importance(model, mutCovariate.table.file, mutCovariate.count.file,
  continuous.features.selected.url.file, z.value = FALSE, mutation.type)
}
\arguments{
\item{model}{Prediction model.}

\item{mutCovariate.table.file}{RDS file, SNV covariates sparse matrix.}

\item{mutCovariate.count.file}{RDS file, SNV response matrix.}

\item{continuous.features.selected.url.file}{Text file containing URLs of SNV continuous features selected for model.}

\item{z.value}{To use z-value for plot or coefficients, default = FALSE.}

\item{mutation.type}{SNV or indel mutation.}
}
\value{
Variable importance plot.
}
\description{
Plot feature importance.
}
