% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Plot manhattan.}
\usage{
plot_manhattan(
  hotspots.file,
  fdr.cutoff = 0.05,
  color.line = "red",
  color.dots = "maroon1"
)
}
\arguments{
\item{hotspots.file}{Hotspots generated.}

\item{fdr.cutoff}{FDR cutoff, default = 0.05.}

\item{color.line}{Color given FDR cutoff, default = red.}

\item{color.dots}{Color hotspots that passed given FDR cutoff, default = maroon1.}
}
\value{
Manhattan figure.
}
\description{
Plot manhattan.
}
