% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.sel.epigenetic.R
\name{stability.sel.epigenetic}
\alias{stability.sel.epigenetic}
\title{100 iterations of LASSO selection for epigenetic features.}
\usage{
stability.sel.epigenetic(x_data, y_data, threshold = 0.75, cores = 1)
}
\arguments{
\item{x_data}{Covariates matrix.}

\item{y_data}{Response matrix.}

\item{threshold}{Frequency cutoff/threshold to determine stable nucleotide contexts, ranges from 0.5 to 1, default = 0.75.}

\item{cores}{Number of cores, default = 1.}
}
\value{
A list containing the frequency of epigenetic features selected through LASSO and the continuous and discrete features that passed the threshold.
}
\description{
100 iterations of LASSO selection for epigenetic features.
}
