% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.sel.nucleotide.R
\name{stability.sel.nucleotide}
\alias{stability.sel.nucleotide}
\title{100 iterations of LASSO selection for nucleotide contexts.}
\usage{
stability.sel.nucleotide(x_data, y_data, threshold = 0.9, cores = 1)
}
\arguments{
\item{x_data}{Covariates matrix.}

\item{y_data}{Response matrix.}

\item{threshold}{Frequency cutoff/threshold to determine stable nucleotide contexts, ranges from 0.5 to 1, default = 0.90.}

\item{cores}{Number of cores, default = 1.}
}
\value{
A list containing the frequency of nucleotide contexts selected through LASSO and the contexts that passed the threshold.
}
\description{
100 iterations of LASSO selection for nucleotide contexts.
}
