% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutAnnotate.R
\name{mutAnnotate}
\alias{mutAnnotate}
\title{Hotspots annotation.}
\usage{
mutAnnotate(hotspots.file, promoter.file = system.file("extdata",
  "Ensembl75.promoters.coding.bed", package = "MutSpot"),
  utr3.file = system.file("extdata", "Ensembl75.3UTR.coding.bed", package =
  "MutSpot"), utr5.file = system.file("extdata", "Ensembl75.5UTR.coding.bed",
  package = "MutSpot"), other.annotations = NULL)
}
\arguments{
\item{hotspots.file}{Hotspots to be annotated.}

\item{promoter.file}{Promoter regions bed file, default file = Ensembl75.promoters.coding.bed.}

\item{utr3.file}{3'UTR regions bed file, default file = Ensembl75.3UTR.coding.bed.}

\item{utr5.file}{5'UTR regions bed file, default file = Ensembl75.5UTR.coding.bed.}

\item{other.annotations}{Text file containing URLs of additional regions to be annotated, default = NULL.}
}
\value{
Annotated hotspots.
}
\description{
Hotspots annotation.
}
