% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.indel.get.features.R
\name{mutPredict.indel.get.features}
\alias{mutPredict.indel.get.features}
\title{Mutation hotspot recurrence prediction for indel.}
\usage{
mutPredict.indel.get.features(
  roi,
  continuous.features.indel,
  discrete.features.indel,
  genome.build = "Ch37"
)
}
\arguments{
\item{roi}{Each indel hotspot in GRanges.}

\item{continuous.features.indel}{Full scores list of indel continuous features selected for model.}

\item{discrete.features.indel}{Full scores list of indel discrete features selected for model.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Feature matrix.
}
\description{
Mutation hotspot recurrence prediction for indel.
}
