% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.indel.run.lr.R
\name{mutPredict.indel.run.lr}
\alias{mutPredict.indel.run.lr}
\title{Run mutation recurrence for indel.}
\usage{
mutPredict.indel.run.lr(output.dir, merge.hotspots = TRUE,
  indel.mutations.file, fdr.cutoff = 0.1, color.line = "red",
  color.dots = "maroon1", color.muts = "orange", top.no = 3,
  promoter.file = system.file("extdata", "Ensembl75.promoters.coding.bed",
  package = "MutSpot"), utr3.file = system.file("extdata",
  "Ensembl75.3UTR.coding.bed", package = "MutSpot"),
  utr5.file = system.file("extdata", "Ensembl75.5UTR.coding.bed", package =
  "MutSpot"), other.annotations = NULL, debug = FALSE, cores = 1)
}
\arguments{
\item{output.dir}{Save plot in given output directory.}

\item{merge.hotspots}{To plot overlapping hotspots as 1 hotspot or individual hotspots, default = TRUE.}

\item{indel.mutations.file}{Indel mutations found in region of interest MAF file.}

\item{fdr.cutoff}{FDR cutoff, default = 0.1.}

\item{color.line}{Color given FDR cutoff, default = red.}

\item{color.dots}{Color hotspots that passed given FDR cutoff, default = maroon1.}

\item{color.muts}{Color points, default = orange.}

\item{top.no}{Number of top hotspots to plot, default = 3.}

\item{promoter.file}{Promoter regions bed file, default file = Ensembl75.promoters.coding.bed.}

\item{utr3.file}{3'UTR regions bed file, default file = Ensembl75.3UTR.coding.bed.}

\item{utr5.file}{5'UTR regions bed file, default file = Ensembl75.5UTR.coding.bed.}

\item{other.annotations}{Text file containing URLs of additional regions to be annotated, default = NULL.}

\item{debug}{To delete temporary files or not, default = FALSE.}

\item{cores}{Number of cores, default = 1.}
}
\value{
Dataframe containing predicted hotspots significance.
}
\description{
Run mutation recurrence for indel.
}
