% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPredict.snv.run.lr.R
\name{mutPredict.snv.run.lr}
\alias{mutPredict.snv.run.lr}
\title{Run mutation recurrence for SNV.}
\usage{
mutPredict.snv.run.lr(
  output.dir,
  merge.hotspots = TRUE,
  snv.mutations.file,
  fdr.cutoff = 0.05,
  color.line = "red",
  color.dots = "maroon1",
  color.muts = "orange",
  top.no = 3,
  promoter.file = system.file("extdata", "Ensembl75.promoters.coding.bed", package =
    "MutSpot"),
  utr3.file = system.file("extdata", "Ensembl75.3UTR.coding.bed", package = "MutSpot"),
  utr5.file = system.file("extdata", "Ensembl75.5UTR.coding.bed", package = "MutSpot"),
  other.annotations = NULL,
  debug = FALSE,
  cores = 1,
  genome.build = "Ch37"
)
}
\arguments{
\item{output.dir}{Results directory.}

\item{merge.hotspots}{To plot overlapping hotspots as 1 hotspot or individual hotspots, default = TRUE.}

\item{snv.mutations.file}{SNV mutations found in region of interest MAF file.}

\item{fdr.cutoff}{FDR cutoff, default = 0.05.}

\item{color.line}{Color given FDR cutoff, default = red.}

\item{color.dots}{Color hotspots that passed given FDR cutoff, default = maroon1.}

\item{color.muts}{Color points, default = orange.}

\item{top.no}{Number of top hotspots to plot, default = 3.}

\item{promoter.file}{Promoter regions bed file, depends on genome build, default file = Ensembl75.promoters.coding.bed, Ch37.}

\item{utr3.file}{3'UTR regions bed file, depends on genome build, default file = Ensembl75.3UTR.coding.bed, Ch37.}

\item{utr5.file}{5'UTR regions bed file, depends on genome build, default file = Ensembl75.5UTR.coding.bed, Ch37.}

\item{other.annotations}{Text file containing URLs of additional regions to be annotated, default = NULL.}

\item{debug}{To delete temporary files or not, default = FALSE.}

\item{cores}{Number of cores, default = 1.}

\item{genome.build}{Reference genome build, default = Ch37.}
}
\value{
Dataframe containing predicted hotspots significance.
}
\description{
Run mutation recurrence for SNV.
}
