% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSpot.R
\name{MutSpot}
\alias{MutSpot}
\title{Runs all or selected steps in MutSpot analysis.}
\usage{
MutSpot(run.to = c(1:2, 3.1, 3.2, 4.1, 4.2, 5.1, 5.2, 5.3, 5.4, 5.5, 6:8, 9.1,
  9.2, 9.3), chromosomes = c(1:22, "X"), snv.mutations = NULL,
  indel.mutations = NULL, mask.regions.file = system.file("extdata",
  "mask_regions.RDS", package = "MutSpot"),
  all.sites.file = system.file("extdata", "all_sites.RDS", package =
  "MutSpot"), region.of.interest = NULL, ratio = 1, sample = T,
  cores = 1, cutoff.nucleotide = 0.9, cutoff.nucleotide.new = NULL,
  genomic.features.snv = NULL, genomic.features.indel = NULL,
  genomic.features = NULL, genomic.features.fixed.snv = NULL,
  genomic.features.fixed.indel = NULL, genomic.features.fixed = NULL,
  sample.snv.features = NULL, sample.indel.features = NULL,
  cutoff.features = 0.75, cutoff.features.new.snv = NULL,
  cutoff.features.new.indel = NULL, fit.sparse = FALSE, drop = FALSE,
  min.count.snv = 2, min.count.indel = 2, genome.size = 2533374732,
  hotspots = TRUE, promoter.file = system.file("extdata",
  "Ensembl75.promoters.coding.bed", package = "MutSpot"),
  utr3.file = system.file("extdata", "Ensembl75.3UTR.coding.bed", package =
  "MutSpot"), utr5.file = system.file("extdata", "Ensembl75.5UTR.coding.bed",
  package = "MutSpot"), other.annotations = NULL, fdr.cutoff = 0.1,
  color.line = "red", color.dots = "maroon1", merge.hotspots = TRUE,
  color.muts = "orange", z.value = FALSE, top.no = 3)
}
\arguments{
\item{run.to}{Numeric vector defining which steps to run, default = 1, 2, 3.1, 4.1, 5.1, 5.2, 5.3, 5.4, 5.5, 6, 7, 8, 9.1, 9.2, 9.3.}

\item{chromosomes}{Character vector defining which chromosomes to compute feature matrix on, default = chr1-chrX.}

\item{snv.mutations}{SNV mutations MAF file, default = NULL.}

\item{indel.mutations}{Indel mutations MAF file, default = NULL.}

\item{mask.regions.file}{Regions to mask in genome, for example, non-mappable regions/immunoglobin loci/CDS regions RDS file, default file = mask_regions.RDS.}

\item{all.sites.file}{All sites in whole genome RDS file, default file = all_sites.RDS.}

\item{region.of.interest}{Region of interest bed file, default = NULL.}

\item{ratio}{Sampling ratio, default = 1.}

\item{sample}{To sample for non-mutated sites or to use all sites in region of interest, default = TRUE.}

\item{cores}{Number of cores, default = 1.}

\item{cutoff.nucleotide}{Frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1, default = 0.90.}

\item{cutoff.nucleotide.new}{Updated frequency cutoff/threshold to determine nucleotide contexts used in prediction model, ranges from 0.5 to 1.}

\item{genomic.features.snv}{Text file containing URLs of potential continuous and discrete SNV epigenetic features to select from, default = NULL.}

\item{genomic.features.indel}{Text file containing URLs of potential continuous and discrete indel epigenetic features to select from, default = NULL.}

\item{genomic.features}{Text file containing URLs of potential continuous and discrete SNV and indel epigenetic features to select from, default = NULL.}

\item{genomic.features.fixed.snv}{Text file containing URLs of fixed continuous and discrete SNV epigenetic features, default = NULL.}

\item{genomic.features.fixed.indel}{Text file containing URLs of fixed continuous and discrete indel epigenetic features, default = NULL.}

\item{genomic.features.fixed}{Text file containing URLs of fixed continuous and discrete SNV and indel epigenetic features, default = NULL.}

\item{sample.snv.features}{Text file containing sample specific SNV features, default = NULL.}

\item{sample.indel.features}{Text file containing sample specific indel features, default = NULL.}

\item{cutoff.features}{Frequency cutoff/threshold to determine epigenetic features used in prediction model, ranges from 0.5 to 1, default = 0.75.}

\item{cutoff.features.new.snv}{Updated frequency cutoff/threshold to determine SNV epigenetic features used in prediction model, ranges from 0.5 to 1.}

\item{cutoff.features.new.indel}{Updated frequency cutoff/threshold to determine indel epigenetic features used in prediction model, ranges from 0.5 to 1.}

\item{fit.sparse}{To fit model using glmnet or glm, default = FALSE.}

\item{drop}{To drop insignificant features from fitted model or not, default = FALSE.}

\item{min.count.snv}{Minimum number of mutated samples in each SNV hotspot, default = 2.}

\item{min.count.indel}{Minimum number of mutated samples in each indel hotspot, default = 2.}

\item{genome.size}{Total number of hotspots to run analysis on, default = 2533374732.}

\item{hotspots}{To run hotspot analysis or region-based analysis, default = TRUE.}

\item{promoter.file}{Promoter regions bed file, default file = Ensembl75.promoters.coding.bed}

\item{utr3.file}{3'UTR regions bed file, default file = Ensembl75.3UTR.coding.bed}

\item{utr5.file}{5'UTR regions bed file, default file = Ensembl75.5UTR.coding.bed}

\item{other.annotations}{Text file containing URLs of additional regions to be annotated, default = NULL.}

\item{fdr.cutoff}{FDR cutoff, default = 0.1.}

\item{color.line}{Color given FDR cutoff, default = red.}

\item{color.dots}{Color hotspots that passed given FDR cutoff, default = maroon1.}

\item{merge.hotspots}{To plot overlapping hotspots as 1 hotspot or individual hotspots, default = TRUE.}

\item{color.muts}{Color points, default = orange.}

\item{z.value}{To use z-value for plot or coefficients, default = FALSE.}

\item{top.no}{Number of top hotspots to plot, default = 3.}
}
\value{
Corresponding output from each step in MutSpot analysis.
}
\description{
Runs all or selected steps in MutSpot analysis.
}
