% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot feature importance.}
\usage{
plot_feature_importance(LRmodel, mutCovariate.table.file, mutation.type,
  output.dir, feature.names)
}
\arguments{
\item{LRmodel}{Prediction model.}

\item{mutCovariate.table.file}{RDS file, SNV covariates sparse matrix.}

\item{mutation.type}{SNV or indel mutation.}

\item{output.dir}{Save plot in given output directory.}

\item{feature.names}{Feature names.}
}
\value{
Variable importance plot.
}
\description{
Plot feature importance.
}
